requirements.txt
aiosqlite
discord
mcstatus
ezcord
python-dotenv